requirements.txt
pytest
pytest-pycodestyle
requests
flask
pymodm
Pillow
freezegun
matplotlib
scipy
numpy
pandas
dnspython
